User Parameters to join tabular & spatial data
When managing assets, it is common to have an list of assets that need to be mapped to be serviced. This workspace contains user parameters and uses feature reader, allowing users to input an asset list (excel) and have these assets in a geospatial format (ready to be mapped).

- FeatureReader transformer allow users to input Esri ArcGIS Server Feature Service (usually available in open data websites) URL. To expose the desired attributes, please, specify the attribute names in the Attribute Exposer transformer.
- To join the tabular data to the geospatial data, it was used Feature Joiner Transformer. Users need to specify the common attributes that will match these 2 datasets.
- The output can be a .shp or .gdb file and the user can choose the desired output coordinate system
- With some minor changes, this workflow can be used on FME Server. It will allow the development of an Workspace App that can be used by non FME users in a wide team.

_____________________________________________________________________________________________

Instead of download the geospatial output in your local, you can make it available in your ArcGIS Enterprise Portal using the Writer Esri ArcGIS Portal Feature Service, allowing non GIS users to verify the location of the assets they want to target.