Add files via upload
This training workspace is designed to facilitate the input of an Excel file containing an asset list for spatial analysis of road levels associated with wastewater pipes' manholes. The analysis includes verifying whether the manhole is positioned within 50 meters of a road intersection. Based on intersection and road level data, the operational traffic management requirements and shifts are assigned for work on these assets.
Using the appropriate parameters on FME Flow, this workspace can be used to create a user-friendly workspace app. This enables non-GIS professionals to conduct intricate spatial analyses without delving too deeply into technical details.
It's important to note that this workspace is intended solely for training purposes. The author assumes no responsibility for its use and content. The data utilised comes originally from open data.