Extract Attribute Values from JSON & Send Survey123 Email Notification
Goal: Send an email notification to users following the submission of a Survey123 inspection, indicating that further action is required.
This workspace extracts attribute values from a JSON file, which is generated whenever a new Survey123 is submitted. To accomplish the email notification, the workspace needs to be published on FME Flow with the appropriate automation parameters configured.